GUIs wurden in Klassen ausgelagert, Hauptmenü wurde angelegt
